Moved Footer speed to the MapContext
Moved footer speed to the MapContext so that it could be accessed globally
